GEODE-6807: update with comments

+++ geode-core/src/main/java/org/apache/geode/distributed/internal/DistributionAdvisor.java
@@ -173,7 +173,7 @@ public class DistributionAdvisor {
   /**
    * Profiles version number
    */
-  protected volatile int profilesVersion = 0;
+  protected volatile long profilesVersion = 0;
 
 
   /**
@@ -1319,7 +1319,6 @@ public class DistributionAdvisor {
   // must synchronize when modifying profile array
   private synchronized boolean basicAddProfile(Profile p) {
     // don't add more than once, but replace existing profile
-    // try {
     profilesVersion++;
     int index = indexOfMemberId(p.getId());
     if (index >= 0) {
@@ -1346,7 +1345,7 @@ public class DistributionAdvisor {
    * Perform work of removing the given member from this advisor.
    */
   private synchronized Profile basicRemoveMemberId(ProfileId id) {
-    // try {
+
     int i = indexOfMemberId(id);
     if (i >= 0) {
       profilesVersion++;
@@ -1355,9 +1354,7 @@ public class DistributionAdvisor {
       return profileRemoved;
     } else
       return null;
-    // } finally {
-    // Assert.assertTrue(-1 == indexOfMemberId(id));
-    // }
+
   }
 
   private int indexOfMemberId(ProfileId id) {
